+++ tests/Doubles/ExtendingTestRule.php
@@ -3,8 +3,9 @@
 namespace Tests\OriNette\ObjectMapper\Doubles;
 
 use Orisai\ObjectMapper\Args\Args;
-use Orisai\ObjectMapper\Context\FieldContext;
-use Orisai\ObjectMapper\Context\TypeContext;
+use Orisai\ObjectMapper\Processing\Context\DynamicContext;
+use Orisai\ObjectMapper\Processing\Context\PropertyContext;
+use Orisai\ObjectMapper\Processing\Context\ServicesContext;
 use Orisai\ObjectMapper\Rules\NoArgsRule;
 use Orisai\ObjectMapper\Types\SimpleValueType;
 use Orisai\ObjectMapper\Types\Type;
@@ -18,12 +19,22 @@ final class ExtendingTestRule extends TestRule
 	 * @param mixed $value
 	 * @return mixed
 	 */
-	public function processValue($value, Args $args, FieldContext $context)
+	public function processValue(
+		$value,
+		Args $args,
+		ServicesContext $services,
+		PropertyContext $property,
+		DynamicContext $dynamic
+	)
 	{
 		return $value;
 	}
 
-	public function createType(Args $args, TypeContext $context): Type
+	public function createType(
+		Args $args,
+		ServicesContext $services,
+		DynamicContext $dynamic
+	): Type
 	{
 		return new SimpleValueType('test');
 	}