+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Support for IntelliJ Neon Pro plugin
 
+### Changed
+
+- Require orisai/object-mapper:^0.3.0
+
 ## [0.1.1](https://github.com/orisai/nette-object-mapper/compare/0.1.0...0.1.1) - 2024-06-22
 
 ### Added